Python/OpenAI.Generate/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.10.0
frozenlist==1.4.0
idna==3.4
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter_client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.5
nltk==3.7
numpy==1.25.1
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas==1.4.3
pandas-stubs==2.0.2.230605
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
platformdirs==3.9.1
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pyzmq==23.2.1
regex==2022.8.17
requests==2.31.0
setuptools==65.5.0
six==1.16.0
stack-data==0.5.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
types-pytz==2023.3.0.0
urllib3==2.0.4
wcwidth==0.2.5
yarl==1.9.2
